Influx extractor converts to IDRF instead of Ingestor
